+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Version 64:
 * Simplify buffered_read_stream composed op
 * Simplify ssl teardown composed op
 * Simplify websocket write_op
+* Exemplars are compiled code
 
 --------------------------------------------------------------------------------